+++ polls/tests/test_poll_dates.py
@@ -1,0 +1,63 @@
+import datetime
+
+from django.test import TestCase
+from django.utils import timezone
+from polls.models import Question
+
+
+class QuestionPollTests(TestCase):
+    def test_was_published_recently_with_future_question(self):
+        """
+        was_published_recently() returns False for questions whose pub_date
+        is in the future.
+        """
+        time = timezone.now() + datetime.timedelta(days=30)
+        future_question = Question(pub_date=time)
+        self.assertIs(future_question.was_published_recently(), False)
+
+    def test_was_published_recently_with_old_question(self):
+        """
+        was_published_recently() returns False for questions whose pub_date
+        is older than 1 day.
+        """
+        time = timezone.now() - datetime.timedelta(days=1, seconds=1)
+        old_question = Question(pub_date=time)
+        self.assertIs(old_question.was_published_recently(), False)
+
+    def test_was_published_recently_with_recent_question(self):
+        """
+        was_published_recently() returns True for questions whose pub_date
+        is within the last day.
+        """
+        time = timezone.now() - datetime.timedelta(hours=23,
+                                                   minutes=59,
+                                                   seconds=59)
+        recent_question = Question(pub_date=time)
+        self.assertIs(recent_question.was_published_recently(), True)
+
+    def test_is_published_with_future_pub_date(self):
+        """
+        is_published() returns False for questions whose
+        pub_date is in the future.
+        """
+        future_time = timezone.now() + datetime.timedelta(days=1)
+        future_question = Question(pub_date=future_time)
+        self.assertFalse(future_question.is_published())
+
+    def test_is_published_with_default_pub_date(self):
+        """
+        is_published() returns True for questions whose
+        pub_date is the current date and time (default).
+        """
+        current_time = timezone.now()
+        current_question = Question(pub_date=current_time)
+        self.assertTrue(current_question.is_published())
+
+    def test_is_published_with_past_pub_date(self):
+        """
+        is_published() returns True for questions whose
+        pub_date is in the past.
+        """
+        past_time = timezone.now() - datetime.timedelta(days=1)
+        past_question = Question(pub_date=past_time)
+        self.assertTrue(past_question.is_published())
